requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.16
botocore==1.17.16
braintree==4.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.0.8
django-countries==6.1.2
django-money==1.1
django-paypal==1.0.0
django-storages==1.9.1
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
# lazy-object-proxy==1.5.0 # astroid incompatible
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2-binary==2.8.5
py-moneyed==0.8.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
